+++ test-resources/odata-service-specs/v4/service-mapping.json
@@ -8,10 +8,5 @@
     "directoryName": "test-service",
     "servicePath": "/sap/opu/odata/sap/API_TEST_SRV",
     "npmPackageName": "test-service"
-  },
-  "COMMON_SRV": {
-    "directoryName": "common-service",
-    "servicePath": "/sap/opu/odata/sap/API_COMMON_ENTITY_SRV/",
-    "npmPackageName": "common-service"
   }
 }
